+++ litellm/litellm_core_utils/core_helpers.py
@@ -4,7 +4,6 @@ import os
 from typing import List, Literal, Optional, Tuple
 
 from litellm._logging import verbose_logger
-from litellm.types.utils import FileTypes
 
 
 def map_finish_reason(
@@ -87,21 +86,3 @@ def _get_parent_otel_span_from_kwargs(kwargs: Optional[dict] = None):
             return kwargs["litellm_parent_otel_span"]
     except:
         return None
-
-
-def get_audio_file_name(file_obj: FileTypes) -> str:
-    """
-    Safely get the name of a file-like object or return its string representation.
-
-    Args:
-        file_obj (Any): A file-like object or any other object.
-
-    Returns:
-        str: The name of the file if available, otherwise a string representation of the object.
-    """
-    if hasattr(file_obj, "name"):
-        return getattr(file_obj, "name")
-    elif hasattr(file_obj, "__str__"):
-        return str(file_obj)
-    else:
-        return repr(file_obj)
